Rename command module
