+++ src/commands/generate/command.ts
