+++ src/commands/generate/index.ts
@@ -15,4 +15,4 @@
  * limitations under the License.
  */
 
-export * from "./generateCommand";
+export * from "./command";
